requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
pkg-resources==0.0.0
PySocks==1.7.1
python-dotenv==0.13.0
requests==2.23.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.9
Werkzeug==1.0.1
